pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "skillprobe"
version = "0.1.0"
description = "A/B evaluate AI agent skills and drive continuous improvement"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "litellm>=1.0",
    "jsonschema>=4.20",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "python-frontmatter>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[project.scripts]
skillprobe = "apps.cli.main:cli"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
